+++ branches/CacheNT/MapView/Map/RMPixel.c
@@ -27,7 +27,7 @@
 
 #include "RMPixel.h"
 
-CGPoint RMScaleCGPointAboutPoint(CGPoint point, float factor, CGPoint pivot)
+CGPoint RMScaleCGPointAboutPoint(CGPoint point, double factor, CGPoint pivot)
 {
 	point.x = (point.x - pivot.x) * factor + pivot.x;
 	point.y = (point.y - pivot.y) * factor + pivot.y;
@@ -35,7 +35,7 @@ CGPoint RMScaleCGPointAboutPoint(CGPoint point, float factor, CGPoint pivot)
 	return point;
 }
 
-CGRect RMScaleCGRectAboutPoint(CGRect rect, float factor, CGPoint pivot)
+CGRect RMScaleCGRectAboutPoint(CGRect rect, double factor, CGPoint pivot)
 {
 	rect.origin = RMScaleCGPointAboutPoint(rect.origin, factor, pivot);
 	rect.size.width *= factor;
