+++ branches/CacheNT/MapView/Map/RMPixel.h
@@ -32,8 +32,8 @@
 
 #include <CoreGraphics/CGGeometry.h>
 
-CGPoint RMScaleCGPointAboutPoint(CGPoint point, float factor, CGPoint pivot);
-CGRect RMScaleCGRectAboutPoint(CGRect rect, float factor, CGPoint pivot);
+CGPoint RMScaleCGPointAboutPoint(CGPoint point, double factor, CGPoint pivot);
+CGRect RMScaleCGRectAboutPoint(CGRect rect, double factor, CGPoint pivot);
 CGPoint RMTranslateCGPointBy(CGPoint point, CGSize delta);
 CGRect RMTranslateCGRectBy(CGRect rect, CGSize delta);
