+++ branches/CacheNT/MapView/Map/RMTiledLayerController.h
@@ -57,7 +57,7 @@
 -(void) centerXYPoint: (RMXYPoint) aPoint animate: (BOOL) animate;
 -(void) centerLatLong: (CLLocationCoordinate2D) point animate: (BOOL) animate;
 -(void) dragBy: (CGSize) delta;
--(void) zoomByFactor: (float) zoomFactor near:(CGPoint) center;
+-(void) zoomByFactor: (double) zoomFactor near:(CGPoint) center;
 
 /*
 -(CGPoint) projectMercatorPoint: (MercatorPoint) point;
